backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
pymysql==1.1.0
cryptography==41.0.7
numpy==1.24.3
pandas==2.1.3
opencv-python==4.8.1.78
onnxruntime==1.16.3
boto3==1.29.7
python-dotenv==1.0.0
Pillow==10.1.0


